+++ lib/essentials/archive_environment/domain/archive_admission_exception.dart
@@ -10,6 +10,7 @@ enum ArchiveAdmissionFailure {
   testApplicationSupportRoot,
   missingMarker,
   nonEmptyUnmarkedArchive,
+  legacyTesterInspectionFailed,
   malformedMarker,
   markerCreationFailed,
   markerFormatMismatch,